requirements.txt
asttokens==2.4.1
attrs==24.2.0
autopep8==2.3.1
backcall==0.2.0
beautifulsoup4==4.12.3
black==24.8.0
bleach==6.1.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.6.1
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.20.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-PyMongo==2.3.0
idna==3.10
importlib_metadata==8.5.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.3.6
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
Pygments==2.18.0
pymongo==4.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
tinycss2==1.3.0
tomli==2.0.1
tornado==6.4.1
traitlets==5.14.3
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
yapf==0.40.2
yarg==0.1.9
zipp==3.20.2
